requirements/base.txt
-e git+https://github.com/django-oscar/django-oscar-accounts.git#egg=django-oscar-accounts
django-oscar-api==1.0.4
# Django
Django==1.9.5

# Environmental var configuration
django-configurations==1.0

# For the persistence stores
psycopg2==2.6.1
dj-database-url==0.4.0

# Models
django-model-utils==2.4

# Rest apis
djangorestframework==3.3.3
Markdown==2.6.6
django-filter==0.13.0

# Static files
whitenoise==3.0

# Media files
boto==2.42.0
# django_unique_upload==0.2.1

# Image manipulation
django-imagekit==3.3

# For asynchronous queuing
django-rq==0.9.0
rq-dashboard==0.3.7
rq-scheduler==0.7.0

# Time zones support
pytz==2016.3

### updated requirements

arrow==0.7.0

Babel==2.3.4
binaryornot==0.4.0
blessings==1.6

cffi==1.6.0
chardet==2.3.0
click==6.6
cookiecutter==1.4.0
cryptography==1.3.1

decorator==4.0.9
defusedxml==0.4.1
dj-database-url==0.4.0
django-allauth==0.25.2
django-autoslug==1.9.3
django-cities==0.4.2
django-extensions==1.6.6
django-phonenumber-field==1.1.0
django-rest-auth==0.7.0
django-storages-redux==1.2.3
django-timedeltafield==0.7.10
django-push-notifications==1.4.1

fake-factory==0.5.7
future==0.15.2

greenlet==0.4.9

idna==2.1
ipython-genutils==0.1.0

Jinja2==2.8
jinja2-time==0.1.0

livereload==2.4.1

MarkupSafe==0.23
mccabe==0.4.0
mkdocs-bootstrap==0.1.1
mkdocs-bootswatch==0.4.0

nose==1.3.7

oauthlib==1.1.1

paramiko==2.0.0
pbr==1.9.1
pep8==1.7.0
pexpect==4.0.1
phonenumberslite==7.3.1
pickleshare==0.7.2
Pillow==3.2.0
poyo==0.2.0
progressbar2==3.6.2
psycopg2==2.6.1
ptyprocess==0.5.1
pyasn1==0.1.9
pycparser==2.14
pyflakes==1.0.0
python-dateutil==2.5.3
python3-openid==3.0.10
PyYAML==3.11

redis==2.10.5
requests==2.10.0
requests-oauthlib==0.6.1
rq==0.5.6

simplegeneric==0.8.1
six==1.10.0

tornado==4.3
traitlets==4.2.1

Unidecode==0.4.19

whichcraft==0.3.1
Werkzeug==0.11.9
drfdocs==0.0.9
django-hstore==1.4.2
django-cors-headers==1.1.0
django-fsm==2.3.0
django-grappelli==2.8.1
pinax-likes==2.0.4
elasticsearch==2.3.0
drf-haystack==1.5.6
drf-extensions==0.2.8
python-magic==0.4.11
django-reversion==1.10.2
pinax-referrals==2.1.0
djangorestframework-hstore==1.3
jsonschema==2.5.1
django-jsonfield==1.0.0
django-activity-stream==0.6.2
django-environ==0.4.0
django-extra-fields==0.7
django-ses==0.8.0
django-field-history==0.5.0
django-review==1.9.5
django-select-multiple-field==0.4.2
-e git+https://github.com/django-haystack/django-haystack.git#egg=django-haystack
